backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
pydantic-settings==2.3.4
passlib==1.7.4
PyJWT==2.8.0
